requirements.txt
fastapi==0.95.1
firebase_admin==6.1.0
protobuf==4.22.3
pydantic==1.10.7
uvicorn==0.21.1
bcrypt
python-decouple
slowapi
aioredis